what is wrong with this source
I think this finally fixes all the errors. lol
