+++ EROverlay.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.32106.194
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ER Overlay", "EROverlay.vcxproj", "{18EC0EBF-58D4-4F9B-B6E0-9E1C8B7D33CF}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "EldenRing_Overlay", "EROverlay.vcxproj", "{18EC0EBF-58D4-4F9B-B6E0-9E1C8B7D33CF}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "EldenRing_VisualLauncher", "GUI Launcher\EldenRing_VisualLauncher.vcxproj", "{8DA67571-375F-42A3-97B2-76BD245BCFC0}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "EldenRing_Launcher", "Launcher\EldenRing_VisualLauncher.vcxproj", "{8DA67571-375F-42A3-97B2-76BD245BCFC0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
