--- GUI Launcher/EldenRing_VisualLauncher.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31205.134
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "EldenRing_VisualLauncher", "EldenRing_VisualLauncher.vcxproj", "{8DA67571-375F-42A3-97B2-76BD245BCFC0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8DA67571-375F-42A3-97B2-76BD245BCFC0}.Debug|x64.ActiveCfg = Debug|x64
-		{8DA67571-375F-42A3-97B2-76BD245BCFC0}.Debug|x64.Build.0 = Debug|x64
-		{8DA67571-375F-42A3-97B2-76BD245BCFC0}.Debug|x86.ActiveCfg = Debug|Win32
-		{8DA67571-375F-42A3-97B2-76BD245BCFC0}.Debug|x86.Build.0 = Debug|Win32
-		{8DA67571-375F-42A3-97B2-76BD245BCFC0}.Release|x64.ActiveCfg = Release|x64
-		{8DA67571-375F-42A3-97B2-76BD245BCFC0}.Release|x64.Build.0 = Release|x64
-		{8DA67571-375F-42A3-97B2-76BD245BCFC0}.Release|x86.ActiveCfg = Release|Win32
-		{8DA67571-375F-42A3-97B2-76BD245BCFC0}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {77C49757-B386-4320-AC29-5EC68E8DB7F0}
-	EndGlobalSection
-EndGlobal
