+++ Overlay/dllmain.cpp
