+++ Overlay/include/D3DRenderer.hpp
