+++ Overlay/src/Common.cpp
