+++ Overlay/src/D3DRenderer.cpp
